requirements.txt
streamlit
pandas
psutil
python-dotenv
watchdog
